Watch YouTube videos directly through the app
These particular edits have been made to allow viewing YouTube videos directly without
having to download them.

HOWEVER, it will only do this through the app itself and not on the splash screen.
Also please note that there is a problem with playing most videos because of DRM
and/or some creators don't allow video embeddings.

As such, this branch will not be merged into the main one.